almost final

+++ Project/src/main/java/gr/ntua/ece/softeng17b/data/Client.java
@@ -19,7 +19,7 @@ public class Client extends User{
 
     // GETTERS
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -87,7 +87,7 @@ public class Client extends User{
 
     // SETTERS
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
